conditional dashPlugin
