+++ docs/index.html
@@ -24,6 +24,14 @@
 
       player = cld.videoPlayer('player');
 
+      player.source(
+        'sea_turtle',
+        {
+          sourceTypes: ['dash/vp9'],
+          info: {title: 'MPEG-DASH vp9'}
+        }
+      );
+
     }, false);
   </script>
 
@@ -38,7 +46,6 @@
       controls
       muted
       class="cld-video-player cld-video-player-skin-dark cld-fluid"
-      data-cld-source='{ "publicId": "sea_turtle", "info": { "title": "Sea turtle", "subtitle": "Short movie about a sea turtle" } }'
     ></video>
 
     <p class="mt-4">
